refactor(deadcode): trim fleet and process exports
